SmartTravelIndia 2/requirements.txt
streamlit>=1.24.0
streamlit
pandas
pillow
